Refactor PayInvoice component
Refactor PayInvoice.tsx into smaller components for better organization and maintainability.
